+++ tests/PHPStan/Analyser/data/curl_getinfo_7.3.php
@@ -9,8 +9,8 @@ class Foo {
 	public function bar()
 	{
 		$handle = new CurlHandle();
-		assertType('float', curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_DOWNLOAD_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_UPLOAD_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_DOWNLOAD_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_CONTENT_LENGTH_UPLOAD_T));
 		assertType('int', curl_getinfo($handle, CURLINFO_HTTP_VERSION));
 		assertType('string', curl_getinfo($handle, CURLINFO_PROTOCOL));
 		assertType('int', curl_getinfo($handle, CURLINFO_PROXY_SSL_VERIFYRESULT));
@@ -19,13 +19,13 @@ class Foo {
 		assertType('int', curl_getinfo($handle, CURLINFO_SIZE_UPLOAD_T));
 		assertType('int', curl_getinfo($handle, CURLINFO_SPEED_DOWNLOAD_T));
 		assertType('int', curl_getinfo($handle, CURLINFO_SPEED_UPLOAD_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_APPCONNECT_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_CONNECT_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_FILETIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_NAMELOOKUP_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_PRETRANSFER_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_REDIRECT_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_STARTTRANSFER_TIME_T));
-		assertType('float', curl_getinfo($handle, CURLINFO_TOTAL_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_APPCONNECT_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_CONNECT_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_FILETIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_NAMELOOKUP_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_PRETRANSFER_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_REDIRECT_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_STARTTRANSFER_TIME_T));
+		assertType('int', curl_getinfo($handle, CURLINFO_TOTAL_TIME_T));
 	}
 }
